requirements.txt
beautifulsoup4~=4.5.3
elasticsearch~=5.1.0
cassandra-driver~=3.7.1
cqlsh~=5.0.3
plotly==1.12.12
numpy==1.11.2
tweepy==3.5.0
nltk==3.2.2
redis==2.10.5
networkx==1.11
matplotlib==2.0.0
scipy==0.18.1